+++ BusinessAssistant-mydata/src/test/java/com/businessassistantbcn/mydata/service/UserSearchesServiceTest.java
@@ -26,7 +26,7 @@ import com.businessassistantbcn.mydata.dto.GenericSearchesResultDto;
 import com.businessassistantbcn.mydata.dto.SaveSearchRequestDto;
 import com.businessassistantbcn.mydata.dto.SaveSearchResponseDto;
 import com.businessassistantbcn.mydata.dto.SearchResultsDto;
-import com.businessassistantbcn.mydata.entities.UserSearch;
+import com.businessassistantbcn.mydata.entity.UserSearch;
 import com.businessassistantbcn.mydata.helper.JsonHelper;
 import com.businessassistantbcn.mydata.repository.UserSearchesRepository;
 import com.fasterxml.jackson.databind.JsonNode;
